Additional GOverviewMapControl functionality
Expand options for adding an overview map with GOverviewMapControl.
In addition to passing a boolean value for the :overview_map option
of GMap#control_init, a hash with :hide and :size options can also
be used.

To have the overview map initialized in the closed state, set the
:hide option to true.

To override the default size of the overview map, set the :size
option to a GSize object.